move multiple agents example into browser dir
